MeshNode::draw - Fixed data conversion warning

+++ src/vk_types.cpp
@@ -13,7 +13,7 @@ void MeshNode::draw(const glm::mat4& topMatrix, DrawContext& ctx)
     {
         RenderObject newRenderObject;
         newRenderObject.startIndex = surface.startIndex;
-        newRenderObject.indicesCount = surface.count;
+        newRenderObject.indicesCount = (uint32_t)surface.count;
         newRenderObject.indexBuffer = meshAsset->meshBuffers.indexBuffer.buffer;
         newRenderObject.material = &surface.material->data;
         newRenderObject.transform = nodeMatrix;
